Refactor: 할일페이지 리팩토링
